Fixed schema load_data

+++ app/models/caboose/utilities/schema.rb
@@ -121,6 +121,7 @@ class Caboose::Utilities::Schema
     return if self.removed_columns.nil?
     c = ActiveRecord::Base.connection
     self.removed_columns.each do |model, columns|
+      next if !c.table_exists?(model.table_name)
       columns.each do |col|
         c.remove_column model.table_name, col if c.column_exists?(model.table_name, col)
       end
